Make everything compatible with SQLite3 in testing and Postgres in dev/prod envs

+++ schema.sql
@@ -4,7 +4,7 @@ CREATE TABLE credentials(
 );
 
 CREATE TABLE users(
-    user_id     INT PRIMARY KEY UNIQUE NOT NULL,
+    user_id     BIGINT PRIMARY KEY UNIQUE NOT NULL,
     name        TEXT NOT NULL,
     email       TEXT NOT NULL
 );
@@ -20,7 +20,7 @@ CREATE TABLE map_cache(
 );
 
 CREATE TABLE location_history(
-    user_id     INT NOT NULL,
+    user_id     BIGINT NOT NULL,
     place_id    TEXT NOT NULL,
     timestamp   REAL NOT NULL,
     FOREIGN KEY(user_id) REFERENCES users(user_id),
